feat: v2.0.50 — shelf improvements, book cover UI, persistence fix
- Fix Android book persistence: libgen downloads now save to persistent
  library folder instead of temp dir (same as Gutenberg downloader)
- Fix shelf bookCount showing 0: add #[serde(rename_all = camelCase)]
  to Shelf model so book_count serializes correctly to frontend
- Redesign ShelfGrid: premium UI with book cover stacks, per-shelf cover
  loading, color-coded cards, hover effects, add-shelf inline card
- Add shelf suggestion on import: generateShelfSuggestions now takes
  full Book objects (fetched via new get_books_by_paths command) so it
  can group by series metadata (Harry Potter, One Piece, etc.) in
  addition to folder structure
- Add get_books_by_paths backend command: efficiently fetches books by
  file paths with authors, tags and manga series name resolution
- SmartShelfSuggestionDialog: use getBooksByPaths instead of getBooks()
  to avoid scanning entire library and fix path matching

+++ src-tauri/src/services/format_adapter.rs
@@ -126,7 +126,7 @@ pub struct BookMetadata {
     pub tags: Vec<String>,
     pub series: Option<String>,
     pub series_index: Option<f32>,
-    pub rating: Option<i32>,
+    pub rating: Option<f64>,
     pub file_format: String,
     pub file_size: u64,
     pub page_count: Option<u32>,
